devices: serial: implement out_timestamp for all platforms
This feature has nothing specific to Windows in its implementation, so
pull it up to the portable part of the Serial device.

The unit test is slightly modified to avoid the need for the regex
crate. This was the only use of regex in devices, so it can be dropped
from Cargo.toml.

BUG=None
TEST=cargo test -p devices serial
TEST=tools/run_tests --platform=mingw64

Change-Id: I3073e90d71bda3c9f6da1df3a3b657e62e087302
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4126251
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>