media: cros-codecs: remove Picture type
The Picture type is a complex generic type which purpose was to hold
both the input and output data of a stateless decoding job. It has
several drawbacks:

- While it maps nicely to VAAPI for which the output of a decoding job
  is known at the time we submit it, it won't work well with V4L2
  stateless which does not allow us to know which CAPTURE buffer will be
  used in advance.
- Before the decoding job is actually submitted, Picture objects are in
  a state where their output is not available, which makes some of its
  methods panic and allows for inconsistent state in the code. Basically
  users of Picture objects are expecting that it is either in a
  pre-submit or a post-submit state, which is dangerous.
- It's a bit complex to understand (arguably opinionated).

There is a clear separation between the pre-submit time, for which only
the input data is available, and the post-submit time, for which we are
only interested in the decoded output handle. By stopping using Picture
we can make sure that only the relevant data is available at both of
these times, and that it is valid - this removes quite a few unwrap()s
and Options and makes the code safer overall.

H.264's DPB was relying on the Picture type as it needs both the frame
information and its decoded handle, so we replace that by a dedicated
DpbEntry type.

This change strips almost 350 LoCs and ~24KB from the binary in a
release build as we now rely less on generics and monomorphization.

Since this change touches to the core of the stateless decoders, it is
necessarily big and a bit all over the place. However the net loss in
LoCs is non-negligible and the code complexity gets reduced quite a bit.

BUG=b:214478588
TEST=`cargo test --features vaapi -p cros-codecs` passes on brya.

Change-Id: Iab735905a7ae5c7285a331ee362e60dbb8f4eea0
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4164537
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>