media: cros-codecs: vaapi: move required data to GenericBackendHandle so it can clean itself up
The destructor for BackendHandle had to take ownership of its
GenericBackendHandle in its destructor in order to return the underlying
surface to the surface pool. This means the handle had to be put behind
an Option that is always valid, but requires us to perform unwrap()s
when accessing it anyway.

By delegating the surface return to GenericBackendHandle, we can avoid
this Option and thus make the code a bit more sound.

BUG=b:214478588
TEST=cargo test --features vaapi -p cros-codecs

Change-Id: I590172b69decac3250bb9e85a67d959b1eca83f8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4164543
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>