p9: handle Tframe body decode failure separately from header
If the header of an incoming message can be decoded, but the body of
that message is invalid in some way (such as a string containing byte
sequences that are not valid UTF-8), the server should respond by
sending the Rerror message rather than closing the whole connection.

If the header cannot be decoded (for example, if the specified size is
too small), it is not possible to resynchronize the byte stream, so the
connection will still be closed in this case.

BUG=b:243123191
TEST=(cd common/p9; cargo test)
TEST=touch /mnt/chromeos/MyFiles/Downloads/$'\351'

Change-Id: I5f10e357a9b1734497bc2d2d8b5de544f86604ae
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4182888
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>