windows: set stdin to raw mode while running a VM
This makes the terminal behave appropriately when using a serial console
connected to stdin/stdout. In particular, there is no extra buffering of
input, and echo is disabled so input is not printed twice.

Attempting to set the terminal mode when running without an attached
console (e.g. within a GUI app where the standard handles are not
connected) will fail silently, which is fine.

BUG=b:264428954
TEST=Use bash running in crosvm from Windows command prompt

Change-Id: I2a267748401ea7d792196f5722c817ba22e1707c
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4200972
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>