crosvm: unix: refactor the interface of jail_helpers
We should use less Option to keep the api simple. But keep the current
interface of `simple_jail()` as is because it is simple enough and fits
other interface easily (i.e. VirtioDeviceStub and other interface expect
to accept `Option<Minijail>`.)

This is the first step of the refactoring jail_helpers to support
injecting /dev/userfaultfd to each device process.

The new chroot path (`src`) for fs_device is already validated when it
is parsed at `SharedDir::from_str()`. But checking the directory twice
won't add much overhead.

BUG=b:266642532
TEST=cargo build

Change-Id: I55f837b4da3df51f878108b0ed3d3aba1a0dd823
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4211151
Commit-Queue: Shin Kawamura <kawasin@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>