crosvm: unix: do not set uid/gid if uid/gid_map are specified
If `ugid_map` is specified but the uid and gid of current user does not
match the map, `setresgid(2)` fails and crosvm fails to boot.

Currently there is not use case when `bind_mounts` is enabled and
`ugid_map` is specified. But this is for the future developers who will
not be bothered by this bug.

BUG=b:266642532
TEST=manual test

Change-Id: I60e1ee18d09ce1ffe05a2a57ea85ba616eb83ccf
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4223603
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>