devices: fix virtio-input not interrupting statusq.
virtio-input has a queue processing thread that handles multiple queues.
Unfortunately the thread collects all of the `needs_interrupt` booleans,
together and then only sends an interrupt for one of the queues (the
eventq). This means that the statusq will never receive its interrupt.
This means we never tell the driver that we have more buffers for it,
and so the device starves and stops sending us guest -> host input
events.

This CL splits up the interrupt tracking so that we interrupt the right
queues as needed.

BUG=b:268572174
TEST=ran emulator with a workload that uses the queue & verified that it
doesn't stall out anymore.

Change-Id: I8a5eb09caff5186a111450da29fcdf0a67087161
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4240542
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Noah Gold <nkgold@google.com>