devices: vhost: vsock: move VhostVsockConfig to the vsock module
We want to use a single option to specify vsock devices on both Linux
and Windows, thus we need a common structures that can be deserialized
from the command-line on both platforms. VhostVsockConfig can fulfill
this purpose if the vhost device parameter is made available on Linux
only, so move it to the virtio/vsock module and rename it to just
VsockConfig.

Since virtio/vsock contains the Windows-only implementation of the
non-vhost vsock device, the visibility of this module needs to be
tuned a bit.

BUG=b:217480043
TEST=cargo build

Change-Id: I4d4a0ef557eded3c6322d213e8c553c573e57b5e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4253289
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>