crosvm: add --vsock command-line option and config-file parameter.
Add a `--vsock` command-line option that is designed to replace the
`--cid`, `--vhost-vsock-device` and `--vhost-vsock-fd` parameters into a
single one. This parameter can be used on Windows as well as Linux, and
on the latter users can also specify the vhost-vsock device to use.

This parameter is the preferred way to add vsock devices, and thus is
also enabled for configuration files. Also update the book's
instructions in lockstep.

BUG=b:255223604
TEST=run crosvm with `--vsock cid=42` option and ensure the guest device
is working using socat to communicate with the host.

Change-Id: I9515922415e5a55289a9e21d5f2f247a30fadc74
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4253292
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>