crosvm: add vsock support for the devices command
Add support for spawning one or several vhost-user devices using the
`devices` command. The syntax is identical to the `--vsock` option of
`crosvm run`, plus the mandatory argument to the vhost-user socket.

This CL mostly reorganizes the seccomp policies of the vhost-vsock device,
since they were not already properly split.

BUG=b:255223604
BUG=b:217480043
BUG=b:217480278
TEST="crosvm devices --vsock vhost=/tmp/vsock-socket,cid=42" results in
a working vhost-vsock device on Linux after a VMM connects to the socket
with the `--vhost-user-vsock=/tmp/vsock-socket` option.
TEST="crosvm devices --vsock vhost=0000:00:10.0,cid=42" results in a
working vhost-vsock device in a guest VM, which another VM can connect
to and use.

Change-Id: Ic2d7a7020dbfb0b3f2971dea361b69241ad87e8c
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4253293
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>