e2e_tests: Do not capture crosvm stdout
This was useful since the libtest harness could not capture stdout
for subprocesses and we would get very noisy test results.

However, with nextest, each test is run in a separate process and
will be hidden unless the test fails.

The test log will then contain an host and guest messages interleaved,
which greatly helps debugging.

Also switches the main serial output from syslog (which prints to
stderr) to stdout.

BUG=None
TEST=None

Change-Id: Ie3b377aeaef3ef575f9232430ac7392c83060d98
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4256511
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>