devices: virtio-balloon: replace Tube::try_clone() usage
The "failable update" option relied on being able to clone the Tube and
leave an instance of it in the main Balloon object while the worker
thread also had a clone of it. To avoid the need for this, create an
extra async function to monitor a queue of pending adjusted responses
that can be written from the write_config() function, along with an
Event to trigger the async function to check the queue and send Adjusted
responses on the command tube.

This also means that all command tube senders can be migrated to the
async version of the function, which is now identical between unix and
win versions and can be moved out of sys.

BUG=b:222588331
TEST=tools/presubmit --all

Change-Id: I1ec384087b69fc28314ff66bff4b0a9560eaba9f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4265534
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>